requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular